Split code in multiple files & `WebRtcPeerCore` base class
